docker/app/requirements.txt
fastapi
gunicorn
uvicorn[standard]
celery[redis]
redis
psycopg2-binary
cryptography
